pyproject.toml
[tool.poetry]
name = "warnings-dropoff-analysis"
version = "0.1.0"
description = ""
authors = ["samuelebortolotti <samuele.bortolotti@hotmail.it>"]

[tool.poetry.dependencies]
python = "^3.8"
pymongo = "^3.11.4"
argparse = "^1.4.0"
pathlib = "^1.0.1"
typing = "^3.7.4"
python-dotenv = "^0.17.1"
bson = "^0.5.10"
utils = "^1.0.1"
pandas = "^1.2.4"
matplotlib = "^3.4.2"
pdoc3 = "^0.9.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pylint = "^2.8.2"
autopep8 = "^1.5.7"
Faker = "^8.1.3"
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
